+++ NplotCompact/LabelPointPlot.cs
@@ -265,7 +265,7 @@ namespace NPlot
 				try
 				{
 					PointD pt = data[i];
-					if ( !Double.IsNaN(pt.X) && !Double.IsNaN(pt.Y) )
+					if ( !Single.IsNaN(pt.X) && !Single.IsNaN(pt.Y) )
 					{
 						PointF xPos = xAxis.WorldToPhysical( pt.X, false);
 						PointF yPos = yAxis.WorldToPhysical( pt.Y, false);
